#300 Refactored ComponentData a bit
