--- bb-aem-64/src/main/java/com/cognifide/qa/bb/aem/core/component/actions/AbstractComponentActionData.java
@@ -1,54 +1,0 @@
-/*-
- * #%L
- * Bobcat
- * %%
- * Copyright (C) 2018 Cognifide Ltd.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package com.cognifide.qa.bb.aem.core.component.actions;
-
-import com.cognifide.qa.bb.api.actions.ActionData;
-
-/**
- *
- */
-public class AbstractComponentActionData implements ActionData {
-
-  private String componentPath;
-
-  private String componentName;
-
-  private int componentOrder;
-
-  public AbstractComponentActionData(String componentPath, String componentName,
-      int componentOrder) {
-    this.componentPath = componentPath;
-    this.componentName = componentName;
-    this.componentOrder = componentOrder;
-  }
-
-  public String getComponentPath() {
-    return componentPath;
-  }
-
-  public String getComponentName() {
-    return componentName;
-  }
-
-  public int getComponentOrder() {
-    return componentOrder;
-  }
-
-}
