+++ bb-aem-64/src/main/java/com/cognifide/qa/bb/aem/core/component/actions/ComponentData.java
@@ -19,13 +19,36 @@
  */
 package com.cognifide.qa.bb.aem.core.component.actions;
 
+import com.cognifide.qa.bb.api.actions.ActionData;
+
 /**
  *
  */
-public class EditComponentData extends AbstractComponentActionData {
+public class ComponentData implements ActionData {
 
-  public EditComponentData(String componentPath, String componentName, int componentOrder) {
-    super(componentPath, componentName, componentOrder);
+  private String componentPath;
+
+  private String componentName;
+
+  private int componentOrder;
+
+  public ComponentData(String componentPath, String componentName,
+      int componentOrder) {
+    this.componentPath = componentPath;
+    this.componentName = componentName;
+    this.componentOrder = componentOrder;
+  }
+
+  public String getComponentPath() {
+    return componentPath;
+  }
+
+  public String getComponentName() {
+    return componentName;
+  }
+
+  public int getComponentOrder() {
+    return componentOrder;
   }
 
 }
