+++ bb-aem-64/src/main/java/com/cognifide/qa/bb/aem/core/component/actions/EditComponent.java
@@ -19,19 +19,22 @@
  */
 package com.cognifide.qa.bb.aem.core.component.actions;
 
-import com.cognifide.qa.bb.api.actions.ActionWithData;
 import com.cognifide.qa.bb.aem.core.component.toolbar.CommonToolbarOptions;
 import com.cognifide.qa.bb.aem.core.component.toolbar.ComponentToolbar;
 import com.cognifide.qa.bb.aem.core.sidepanel.internal.SidePanel;
 import com.cognifide.qa.bb.aem.core.sidepanel.internal.SidePanelTabs;
+import com.cognifide.qa.bb.api.actions.ActionWithData;
 import com.cognifide.qa.bb.qualifier.FindPageObject;
 import com.cognifide.qa.bb.qualifier.Global;
 import com.cognifide.qa.bb.qualifier.PageObject;
 
 import io.qameta.allure.Step;
 
+/**
+ * An {@link ActionWithData} that edits the component described in {@link ComponentData}
+ */
 @PageObject
-public class EditComponent implements ActionWithData<EditComponentData> {
+public class EditComponent implements ActionWithData<ComponentData> {
 
   @FindPageObject
   private SidePanel sidePanel;
@@ -42,7 +45,7 @@ public class EditComponent implements ActionWithData<EditComponentData> {
 
   @Override
   @Step("Edit component")
-  public void execute(EditComponentData data) {
+  public void execute(ComponentData data) {
     sidePanel.selectTab(SidePanelTabs.CONTENT_TREE.getCssClass());
     sidePanel.selectComponentToEdit(data.getComponentPath(), data.getComponentName(),
         data.getComponentOrder()).click();
